// src/modules/pos-pre-sapi/interfaces/posiciones-presupuestales-sapi.interfaces.ts
export interface IPosPreSapi {
    id?: number;
    numberName: string;
    exercise: number;
    description: string;
    consecutive: string;
    assignedTo: string;
    status: boolean;
    createDocumentUserAt?: string;
    createDateAt?: Date;
    updateDocumentUserAt?: string;
    updateDateAt?: Date;
}